Add UTF-8 BOM to CSV reports
